move SpawnController to World package
