+++ src/main/java/core/levels/LevelManager.java
@@ -5,6 +5,7 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.maps.tiled.TiledMap;
 import com.badlogic.gdx.maps.tiled.TmxMapLoader;
 import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
+import core.world.SpawnController;
 import core.world.WorldEntity;
 import core.world.WorldManager;
 
@@ -46,7 +47,7 @@ public class LevelManager {
     }
 
     // TODO: Configure changes from SpawnController
-    public void spawnEntity(WorldEntity worldEntity, SpawnController spawnController){
+    public void spawnEntity(WorldEntity worldEntity, SpawnController spawnController) {
         spawnController.spawn(worldEntity);
     }
 
@@ -61,6 +62,4 @@ public class LevelManager {
     public void dispose() {
         map.dispose();
     }
-
-
 }
